requirements.txt
# ============================================================
# World News Intelligence Dashboard – Python Dependencies
# Python 3.11+
# ============================================================

# Core
feedparser==6.0.11
requests==2.31.0
python-dateutil==2.9.0
pytz==2024.1
pyyaml==6.0.1

# Storage
duckdb==0.10.3

# Data processing
pandas==2.2.2

# NLP
vaderSentiment==3.3.2
scikit-learn==1.5.0

# Dashboard
streamlit==1.35.0
plotly==5.22.0

# Dev / testing
pytest==8.2.0
pytest-cov==5.0.0
black==24.4.2
ruff==0.4.7
